requirements.txt
streamlit
llama-parse
llama-index
google-generativeai
python-dotenv
nest-asyncio
llama-index-llms-gemini
llama-index-embeddings-gemini
faiss-cpu
numpy
